Change the type of EnumerateeHandler
There was a bug in eneeCheckIfDoneHandle where certain returned exceptions
weren't handled.  In order to fix this, it's necessary to be able to handle
exceptions from the outer iteratee, not just the inner iteratee.  Converting
exception handlers shouldn't be too difficult as an inner iteratee can just be
wrapped in 'return' to lift it to the outer.

This also fixes some bugs in eneeCheckIfDone* where the recovery continuation
wasn't set properly.